moved flink error in the errors package
