+++ internal/pkg/ccloudv2/flink_gateway.go
@@ -2,7 +2,6 @@ package ccloudv2
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/google/uuid"
 
@@ -44,12 +43,12 @@ func (c *FlinkGatewayClient) flinkGatewayApiContext() context.Context {
 
 func (c *FlinkGatewayClient) DeleteStatement(environmentId, statementName, orgId string) error {
 	httpResp, err := c.StatementsSqlV1alpha1Api.DeleteSqlV1alpha1Statement(c.flinkGatewayApiContext(), environmentId, statementName).OrgId(orgId).Execute()
-	return catchFlinkError(err, httpResp)
+	return errors.CatchFlinkError(err, httpResp)
 }
 
 func (c *FlinkGatewayClient) GetStatement(environmentId, statementName, orgId string) (flinkgatewayv1alpha1.SqlV1alpha1Statement, error) {
 	resp, httpResp, err := c.StatementsSqlV1alpha1Api.GetSqlV1alpha1Statement(c.flinkGatewayApiContext(), environmentId, statementName).OrgId(orgId).Execute()
-	return resp, catchFlinkError(err, httpResp)
+	return resp, errors.CatchFlinkError(err, httpResp)
 }
 
 func (c *FlinkGatewayClient) ListStatements(environmentId, orgId, pageToken, computePoolId string) (flinkgatewayv1alpha1.SqlV1alpha1StatementList, error) {
@@ -63,7 +62,7 @@ func (c *FlinkGatewayClient) ListStatements(environmentId, orgId, pageToken, com
 		req = req.PageToken(pageToken)
 	}
 	resp, httpResp, err := req.Execute()
-	return resp, catchFlinkError(err, httpResp)
+	return resp, errors.CatchFlinkError(err, httpResp)
 }
 
 func (c *FlinkGatewayClient) ListAllStatements(environmentId, orgId, computePoolId string) ([]flinkgatewayv1alpha1.SqlV1alpha1Statement, error) {
@@ -98,7 +97,7 @@ func (c *FlinkGatewayClient) CreateStatement(statement, computePoolId, identityP
 		},
 	}
 	resp, httpResp, err := c.StatementsSqlV1alpha1Api.CreateSqlV1alpha1Statement(c.flinkGatewayApiContext(), environmentId).SqlV1alpha1Statement(statementObj).OrgId(orgId).Execute()
-	return resp, catchFlinkError(err, httpResp)
+	return resp, errors.CatchFlinkError(err, httpResp)
 }
 
 func (c *FlinkGatewayClient) GetStatementResults(environmentId, statementId, orgId, pageToken string) (flinkgatewayv1alpha1.SqlV1alpha1StatementResult, error) {
@@ -107,65 +106,10 @@ func (c *FlinkGatewayClient) GetStatementResults(environmentId, statementId, org
 		req = req.PageToken(pageToken)
 	}
 	resp, httpResp, err := req.Execute()
-	return resp, catchFlinkError(err, httpResp)
+	return resp, errors.CatchFlinkError(err, httpResp)
 }
 
 func (c *FlinkGatewayClient) GetExceptions(environmentId, statementId, orgId string) (flinkgatewayv1alpha1.SqlV1alpha1StatementExceptionList, error) {
 	resp, httpResp, err := c.StatementExceptionsSqlV1alpha1Api.GetSqlV1alpha1StatementExceptions(c.flinkGatewayApiContext(), environmentId, statementId).OrgId(orgId).Execute()
-	return resp, catchFlinkError(err, httpResp)
-}
-
-// FlinkError extends the ErrorWithSuggestion with a status code.
-type FlinkError struct {
-	errorMsg       string
-	suggestionsMsg string
-	statusCode     int
-}
-
-func NewFlinkError(errorMsg string, suggestionsMsg string, statusCode int) FlinkError {
-	return FlinkError{
-		errorMsg:       errorMsg,
-		suggestionsMsg: suggestionsMsg,
-		statusCode:     statusCode,
-	}
-}
-
-func (f FlinkError) StatusCode() int {
-	return f.statusCode
-}
-
-func (f FlinkError) GetSuggestionsMsg() string {
-	return f.suggestionsMsg
-}
-
-func (f FlinkError) Error() string {
-	return f.errorMsg
-}
-
-type Coder interface {
-	StatusCode() int
-}
-
-var _ Coder = (*FlinkError)(nil)
-var _ errors.ErrorWithSuggestions = (*FlinkError)(nil)
-
-// Extends error with status code, including suggestion if err type is ErrorWithSuggestion
-func catchFlinkError(err error, r *http.Response) error {
-	if err == nil {
-		return nil
-	}
-	err = errors.CatchCCloudV2Error(err, r)
-	suggestion := ""
-	if suggester, ok := err.(errors.ErrorWithSuggestions); ok {
-		suggestion = suggester.GetSuggestionsMsg()
-	}
-	var statusCode int
-	if r != nil {
-		statusCode = r.StatusCode
-	}
-	return FlinkError{
-		statusCode:     statusCode,
-		errorMsg:       err.Error(),
-		suggestionsMsg: suggestion,
-	}
+	return resp, errors.CatchFlinkError(err, httpResp)
 }
