+++ internal/pkg/ccloudv2/flink_gateway_test.go
@@ -14,64 +14,64 @@ import (
 func TestFlinkErrorCodeWhenErrors(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"errors":[{"detail":"There is an error"}]}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := catchFlinkError(errors.New("Some Error"), res)
+	err := errors.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(FlinkError)
+	flinkError, ok := err.(errors.FlinkError)
 	require.True(t, ok)
-	require.Equal(t, http.StatusMethodNotAllowed, flinkError.statusCode)
+	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
 }
 
 func TestFlinkErrorNil(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"errors":[{"detail":"There is an error"}]}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := catchFlinkError(nil, res)
+	err := errors.CatchFlinkError(nil, res)
 	require.Nil(t, err)
 }
 
 func TestFlinkErrorNilHttpRes(t *testing.T) {
-	err := catchFlinkError(errors.New("Some Error"), nil)
+	err := errors.CatchFlinkError(errors.New("Some Error"), nil)
 	require.Error(t, err)
 
-	flinkError, ok := err.(FlinkError)
+	flinkError, ok := err.(errors.FlinkError)
 	require.True(t, ok)
-	require.Equal(t, 0, flinkError.statusCode)
+	require.Equal(t, 0, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
 }
 
 func TestFlinkErrorCodeWhenErrorMessage(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"message":"unauthorized"}`)), StatusCode: http.StatusUnauthorized}
 
-	err := catchFlinkError(errors.New("Some Error"), res)
+	err := errors.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(FlinkError)
+	flinkError, ok := err.(errors.FlinkError)
 	require.True(t, ok)
-	require.Equal(t, http.StatusUnauthorized, flinkError.statusCode)
+	require.Equal(t, http.StatusUnauthorized, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
 }
 
 func TestFlinkErrorCodeWhenNestedMessage(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"error":{"message":"gateway error"}}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := catchFlinkError(errors.New("Some Error"), res)
+	err := errors.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(FlinkError)
+	flinkError, ok := err.(errors.FlinkError)
 	require.True(t, ok)
-	require.Equal(t, http.StatusMethodNotAllowed, flinkError.statusCode)
+	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
 }
 
 func TestFlinkErrorOnlyStatusCode(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader("")), StatusCode: http.StatusMethodNotAllowed}
 
-	err := catchFlinkError(errors.New("Some Error"), res)
+	err := errors.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(FlinkError)
+	flinkError, ok := err.(errors.FlinkError)
 	require.True(t, ok)
-	require.Equal(t, http.StatusMethodNotAllowed, flinkError.statusCode)
+	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
 }
