+++ internal/pkg/flink/internal/controller/statement_controller_test.go
@@ -2,6 +2,7 @@ package controller
 
 import (
 	"context"
+	"net/http"
 	"testing"
 	"time"
 
@@ -105,7 +106,7 @@ func (s *StatementControllerTestSuite) TestExecuteStatement() {
 
 func (s *StatementControllerTestSuite) TestExecuteStatementExitApplicationOnUnauthorizedResponse() {
 	statementToExecute := "select 1;"
-	processedStatementError := types.StatementError{Message: "unauthorized", StatusCode: 401}
+	processedStatementError := types.StatementError{Message: "unauthorized", StatusCode: http.StatusUnauthorized}
 	s.store.EXPECT().ProcessStatement(statementToExecute).Return(nil, &processedStatementError)
 	s.applicationController.EXPECT().ExitApplication()
